refactor: Button -> BaseButton 이름 변경 및 상수파일 분리
