+++ src/common/index.ts
@@ -1,2 +1,2 @@
 export { default as Sidebar } from "./Sidebar/Sidebar";
-export { default as Button } from "./Button/Button";
+export { default as BaseButton } from "./Button/BaseButton";
